+++ docs/release-notes.md
@@ -5,11 +5,12 @@
 
 ### New
 
+- Tray menu has "URL Tester" shortcut.
 - "buy me a coffee" link removed from the tray icon and status bar, because no one is donating, you cheap bastards!
 
 ## 3.4.2
 
-Bug fixed - picker was always defaulting to "chomeless window". Thanks to @danarnold in #20.
+Bug fixed - picker was always defaulting to "chrome-less window". Thanks to @danarnold in #20.
 
 ## 3.4.1